Moved DomainSplitter to kernel package
