+++ src/DGtal/kernel/domains/DomainSplitter.h
@@ -1,4 +1,30 @@
+/**
+ *  This program is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU Lesser General Public License as
+ *  published by the Free Software Foundation, either version 3 of the
+ *  License, or  (at your option) any later version.
+ *
+ *  This program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License
+ *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ **/
+
 #pragma once
+
+/**
+ * @file DomainSplitter.h
+ * @author Bastien DOIGNIES (\c bastien.doignies@liris.cnrs.fr )
+ * LIRIS (CNRS, UMR 5205), University of Lyon, France
+ *
+ * @date 2026/04/29
+ *
+ * This file is part of the DGtal library.
+ */
 
 #include "DGtal/kernel/domains/CDomain.h"
 #include <vector>
