+++ tests/projects/CMakeLists.txt
@@ -2,6 +2,7 @@ project(SystemWideCompile)
 cmake_minimum_required(VERSION 3.20)
 
 set(CMAKE_CXX_STANDARD 20)
+set(CMAKE_PREFIX_PATH "/home/pattoune/lib/DGtal")
 
 find_package(DGtal REQUIRED)
 add_executable(SystemWideCompile main.cpp)
